Remove double-blind application and library names
As the paper has been accepted, the library and the
application can be reverted to their original names.
Eyrie and Chirper had been changed to, respectively,
Nest and Flitter, to prevent association of the work
with its authors. The reason for that was the double
blind review process. Also, a minor fix was made, as
a non-existing reference was being \cite{}ed.